refactor: 提取轮询逻辑到 useTaskPolling composable
- 新增 composables/useTaskPolling.js
- 封装轮询启动、停止、回调逻辑
- 自动在组件卸载时清理定时器
- App.vue 使用 composable 简化代码
