+++ frontend/src/App.vue
@@ -1,6 +1,7 @@
 <script setup>
 import { ref, reactive, computed } from 'vue'
-import { createTask, getTask } from './api/task'
+import { createTask } from './api/task'
+import { useTaskPolling } from './composables/useTaskPolling'
 import AppHeader from './components/AppHeader.vue'
 import AppFooter from './components/AppFooter.vue'
 import InitialPage from './components/InitialPage.vue'
@@ -10,7 +11,6 @@ import ResultPage from './components/ResultPage.vue'
 const page = ref('initial')
 const url = ref('')
 const downloadOnly = ref(false)
-const taskId = ref(null)
 const taskStatus = ref('pending')
 const currentTab = ref('article')
 const errorMessage = ref('无法处理该视频链接，请检查链接是否正确且可公开访问，或尝试其他视频。')
@@ -31,9 +31,54 @@ const result = reactive({
     transcript: ''
 })
 
-let pollInterval = null
 let lastUrl = ''
 let lastDownloadOnly = false
+
+// 状态映射
+const statusMap = {
+    'pending': { text: '等待处理...', percent: 10, step: '步骤 1/3' },
+    'downloading': { text: '正在下载视频...', percent: 33, step: '步骤 1/3' },
+    'transcribing': { text: '正在转录音频...', percent: 55, step: '步骤 2/3' },
+    'generating': { text: '正在生成内容...', percent: 80, step: '步骤 3/3' },
+    'completed': { text: '处理完成', percent: 100, step: '完成' },
+}
+
+// 轮询
+const { taskId, start: startPolling, stop: stopPolling } = useTaskPolling({
+    onUpdate(data) {
+        taskStatus.value = data.status
+
+        // 渐进更新数据
+        if (data.title) {
+            progress.title = data.title
+            result.title = data.title
+        }
+        if (data.has_video) result.has_video = true
+        if (data.has_audio) result.has_audio = true
+        if (data.transcript) result.transcript = data.transcript
+        if (data.outline) result.outline = data.outline
+        if (data.article) result.article = data.article
+
+        // 更新进度条
+        if (statusMap[data.status]) {
+            const s = statusMap[data.status]
+            progress.text = data.progress || s.text
+            progress.percent = s.percent
+            progress.step = s.step
+        }
+
+        // 自动切换到有内容的tab
+        if (result.transcript && !result.outline && !result.article && currentTab.value !== 'transcript') {
+            currentTab.value = 'transcript'
+        }
+    },
+    onComplete() {
+        currentTab.value = result.article ? 'article' : (result.outline ? 'outline' : 'transcript')
+    },
+    onFailed(data) {
+        errorMessage.value = data.error || '处理失败'
+    }
+})
 
 // 当前内容（用于复制）
 const currentContent = computed(() => {
@@ -44,7 +89,6 @@ const currentContent = computed(() => {
 
 // 方法
 const resetState = () => {
-    taskId.value = null
     taskStatus.value = 'pending'
     currentTab.value = 'article'
     errorMessage.value = '无法处理该视频链接，请检查链接是否正确且可公开访问，或尝试其他视频。'
@@ -53,14 +97,14 @@ const resetState = () => {
 }
 
 const startNew = () => {
-    if (pollInterval) clearInterval(pollInterval)
+    stopPolling()
     resetState()
     url.value = ''
     page.value = 'initial'
 }
 
 const retryTask = () => {
-    if (pollInterval) clearInterval(pollInterval)
+    stopPolling()
     resetState()
     url.value = lastUrl
     downloadOnly.value = lastDownloadOnly
@@ -80,71 +124,10 @@ const submitUrl = async () => {
 
     try {
         const data = await createTask(url.value, lastDownloadOnly)
-        taskId.value = data.task_id
-        startPolling()
+        startPolling(data.task_id)
     } catch (error) {
         errorMessage.value = error.message
         taskStatus.value = 'failed'
-    }
-}
-
-const startPolling = () => {
-    if (pollInterval) clearInterval(pollInterval)
-    pollInterval = setInterval(checkTask, 2000)
-    checkTask()
-}
-
-const checkTask = async () => {
-    if (!taskId.value) return
-
-    try {
-        const data = await getTask(taskId.value)
-
-        // 更新任务状态
-        taskStatus.value = data.status
-
-        // 渐进更新数据
-        if (data.title) {
-            progress.title = data.title
-            result.title = data.title
-        }
-        if (data.has_video) result.has_video = true
-        if (data.has_audio) result.has_audio = true
-        if (data.transcript) result.transcript = data.transcript
-        if (data.outline) result.outline = data.outline
-        if (data.article) result.article = data.article
-
-        // 更新进度条
-        const statusMap = {
-            'pending': { text: '等待处理...', percent: 10, step: '步骤 1/3' },
-            'downloading': { text: '正在下载视频...', percent: 33, step: '步骤 1/3' },
-            'transcribing': { text: '正在转录音频...', percent: 55, step: '步骤 2/3' },
-            'generating': { text: '正在生成内容...', percent: 80, step: '步骤 3/3' },
-            'completed': { text: '处理完成', percent: 100, step: '完成' },
-        }
-
-        if (statusMap[data.status]) {
-            const s = statusMap[data.status]
-            progress.text = data.progress || s.text
-            progress.percent = s.percent
-            progress.step = s.step
-        }
-
-        // 自动切换到有内容的tab
-        if (result.transcript && !result.outline && !result.article && currentTab.value !== 'transcript') {
-            currentTab.value = 'transcript'
-        }
-
-        if (data.status === 'completed') {
-            clearInterval(pollInterval)
-            // 切换到最佳内容tab
-            currentTab.value = result.article ? 'article' : (result.outline ? 'outline' : 'transcript')
-        } else if (data.status === 'failed') {
-            clearInterval(pollInterval)
-            errorMessage.value = data.error || '处理失败'
-        }
-    } catch (error) {
-        console.error('轮询失败:', error)
     }
 }
